Fix bug with DatePicker min and max not updating during cloning
This works around a bug where the minDate and maxDate values passed to DatePicker were not propagating to the calendar component. Once course start was set, that would become the both the min and max for the timeline start during the course  cloning process. The updated date limits were being passed, but because the actual target input's value didn't change, InputHOC was short-circuiting the updates. I don't think it always behaved like this, but I'm not sure when the behavior changed; it might have been from a React update or a related library.

Ultimately, we probably want to get rid of InputHOC altogether and use a more standard way to manage performance for redux-connected inputs.